// types.ts
export type ExamType = 'TNPSC' | 'Bank Exam' | 'Railway' | 'SSC' | (string & {});
// ...
export type TNPSCGroupFilterType = 'Group I' | 'Group II' | 'Group IV' | (string & {});
export type TNPSCStageFilterType = 'PYQ' | 'Prelims' | 'Mains' | 'Full Mock Test' | 'Upload Files' | (string & {});
export type BankFilterType = 'PYQ' | 'Prelims' | 'Mains' | 'Full Mock Test' | 'Upload Files' | 'Sample Questions' | (string & {});
// FIX: Corrected RRB to RRB BANK to match constants
export type BankExamFilterType = 'SBI' | 'IBPS' | 'RRB BANK' | (string & {});
export type RailwayExamFilterType = 'NTPC' | 'Group D' | 'RRB ALP' | (string & {});
export type RailwayStageFilterType = 'PYQ' | 'CBT-1' | 'CBT-2' | 'CBT-2 (Part A)' | 'CBT-2 (Part B)' | 'Part B' | 'Full Mock Test' | 'Upload Files' | 'Sample Questions' | (string & {});
export type SSCExamFilterType = 'CGL' | 'CHSL' | 'JE' | (string & {});
export type SSCStageFilterType = 'PYQ' | 'Tier-1' | 'Tier-2' | 'Paper-I' | 'Paper-II' | 'Full Mock Test' | 'Upload Files' | 'Sample Questions' | (string & {});